+++ src/config/incluster_config.rs
@@ -1,7 +1,7 @@
 use std::env;
 
 use crate::{Result, Error};
-use openssl::x509::X509;
+use reqwest::Certificate;
 
 use crate::config::utils;
 
@@ -31,9 +31,9 @@ pub fn load_token() -> Result<String> {
 }
 
 /// Returns certification from specified path in cluster.
-pub fn load_cert() -> Result<X509> {
+pub fn load_cert() -> Result<Certificate> {
     let ca = utils::data_or_file_with_base64(&None, &Some(SERVICE_CERTFILE))?;
-    X509::from_pem(&ca).map_err(|e| Error::KubeConfig(format!("{}", e)))
+    Certificate::from_pem(&ca).map_err(|e| Error::KubeConfig(format!("{}", e)))
 }
 
 /// Returns the default namespace from specified path in cluster.