+++ src/config/kube_config.rs
@@ -1,11 +1,28 @@
 use std::path::Path;
+
+#[cfg(feature = "native-tls")]
 use openssl::{
     pkcs12::Pkcs12,
     pkey::PKey,
     x509::X509,
 };
+
+use reqwest::{Identity, Certificate};
+
 use crate::{Error, Result};
 use crate::config::apis::{AuthInfo, Cluster, Config, Context};
+
+/// Regardless of tls type, a Certificate Der is always a byte array
+pub struct Der(pub Vec<u8>);
+
+use std::convert::TryFrom;
+impl TryFrom<Der> for Certificate {
+    type Error = Error;
+    fn try_from(val: Der) -> Result<Certificate> {
+        Certificate::from_der(&val.0)
+            .map_err(Error::ReqwestError)
+    }
+}
 
 /// KubeConfigLoader loads current context, cluster, and authentication information.
 #[derive(Clone,Debug)]
@@ -39,7 +56,6 @@ impl KubeConfigLoader {
             .ok_or_else(|| Error::KubeConfig("Unable to load cluster of context".into()))?;
         let user_name = user.as_ref().unwrap_or(&current_context.user);
 
-        // NB: can't have async closures yet
         let mut user_opt = None;
         for named_user in config.auth_infos {
             if &named_user.name == user_name {
@@ -56,20 +72,63 @@ impl KubeConfigLoader {
         })
     }
 
-    pub fn p12(&self, password: &str) -> Result<Pkcs12> {
+    #[cfg(feature="native-tls")]
+    pub fn identity(&self, password: &str) -> Result<Identity> {
         let client_cert = &self.user.load_client_certificate()?;
         let client_key = &self.user.load_client_key()?;
 
-        let x509 = X509::from_pem(&client_cert).map_err(|e| Error::SslError(format!("{}", e)))?;
-        let pkey = PKey::private_key_from_pem(&client_key).map_err(|e| Error::SslError(format!("{}", e)))?;
+        let x509 = X509::from_pem(&client_cert)
+            .map_err(|e| Error::SslError(format!("{}", e)))?;
+        let pkey = PKey::private_key_from_pem(&client_key)
+            .map_err(|e| Error::SslError(format!("{}", e)))?;
 
-        Pkcs12::builder()
+        let p12 = Pkcs12::builder()
             .build(password, "kubeconfig", &pkey, &x509)
+            .map_err(|e| Error::SslError(format!("{}", e)))?;
+
+        let der = p12.to_der()
+            .map_err(|e| Error::SslError(format!("{}", e)))?;
+        Ok(Identity::from_pkcs12_der(&der, password)?)
+    }
+
+    #[cfg(feature="rustls-tls")]
+    pub fn identity(&self, _password: &str) -> Result<Identity> {
+        let client_cert = &self.user.load_client_certificate()?;
+        let client_key = &self.user.load_client_key()?;
+
+        let mut buffer = client_key.clone();
+        buffer.extend_from_slice(client_cert);
+        Identity::from_pem(&buffer.as_slice())
             .map_err(|e| Error::SslError(format!("{}", e)))
     }
 
-    pub fn ca_bundle(&self) -> Option<Result<Vec<X509>>> {
-        let bundle = self.cluster.load_certificate_authority().ok()?;
-        Some(X509::stack_from_pem(&bundle).map_err(|e| Error::SslError(format!("{}", e))))
+    #[cfg(feature="native-tls")]
+    pub fn ca_bundle(&self) -> Result<Vec<Der>> {
+        let bundle = self.cluster.load_certificate_authority()
+            .map_err(|e| Error::SslError(format!("{}", e)))?;
+        let bundle = X509::stack_from_pem(&bundle)
+            .map_err(|e| Error::SslError(format!("{}", e)))?;
+
+        let mut stack = vec![];
+        for ca in bundle {
+            let der = ca.to_der().map_err(|e| Error::SslError(format!("{}", e)))?;
+            stack.push(Der(der))
+        }
+        Ok(stack)
+    }
+
+    #[cfg(feature = "rustls-tls")]
+    pub fn ca_bundle(&self) -> Result<Vec<Der>> {
+        use rustls::internal::pemfile;
+        use std::io::Cursor;
+        let bundle = self.cluster.load_certificate_authority()?;
+        let mut pem = Cursor::new(bundle);
+        pem.set_position(0);
+
+        let mut stack = vec![];
+        for ca in pemfile::certs(&mut pem).map_err(|e| Error::SslError(format!("{:?}", e)))? {
+            stack.push(Der(ca.0))
+        }
+        Ok(stack)
     }
 }