+++ src/oauth2/mod.rs
@@ -4,10 +4,14 @@ use std::path::PathBuf;
 
 use chrono::Utc;
 use crate::{Result, Error};
-use openssl::pkey::{PKey, Private};
-use openssl::sign::Signer;
-use openssl::rsa::Padding;
-use openssl::hash::MessageDigest;
+
+#[cfg(feature = "native-tls")]
+use openssl::{
+    pkey::{PKey, Private},
+    sign::Signer,
+    rsa::Padding,
+    hash::MessageDigest,
+};
 use reqwest::Client;
 use reqwest::header::CONTENT_TYPE;
 //use time::Duration;
@@ -117,6 +121,8 @@ impl CredentialsClient {
             client: Client::new(),
         })
     }
+
+    #[cfg(feature = "native-tls")]
     pub async fn request_token(&self, scopes: &[String]) -> Result<Token> {
         let private_key = PKey::private_key_from_pem(&self.credentials.private_key.as_bytes())
             .map_err(|e| Error::SslError(format!("{}", e)))?;
@@ -146,6 +152,7 @@ impl CredentialsClient {
         Ok(token_response.into_token())
     }
 
+    #[cfg(feature = "native-tls")]
     fn jws_encode(&self, claim: &Claim, header: &Header, key: PKey<Private>) -> Result<String> {
         let encoded_header = self.base64_encode(serde_json::to_string(&header).unwrap().as_bytes());
         let encoded_claims = self.base64_encode(serde_json::to_string(&claim).unwrap().as_bytes());